Add clarification to post-subs purchase "Register for an account"
This was requested by Gill Crew on the subs team - to make it clearer to the user what is that they are signing in _to_.

I haven't made the text of theguardian.com a link, because I fear that would lead users away from completing the password entry. 

---------- Forwarded message ----------
From: Suzanne O'Brien
Date: 7 September 2015 at 11:00
Subject: Text Update - nice to have but not a blocker

...

secondly - Gill has requested an small update to the wording used under Register for an account and would like the link to the guardian to be included.

Think it might be useful to mention theguardian.com in there somewhere there's no mention of where they are actually creating an account for. Would it be possible to change the copy to read:


Some features of your subscription require you to be signed in to theguardian.com before you can make use of them.