Allow codes that last a full year in the CAS code generator tool
Paul Browne (User Help) advised that the old CAS lookup tool allowed us to generate a 52-week token, but the new version only goes up to 13 weeks. Is it possible for this to be extended? This would be useful for anyone having trouble getting access to the 6-month Galaxy Gifts offer for the Guardian app or for testing.