Committing for backup
Use new membership-common API to alter the start date of the 6-for-6 rate plan charge in a different way, such that the contract effective date can be set to 'now' and be consistent with all other subs. Also the + - one day (aka 'starting the six for six on a Thursday') is no longer necessary as the fulfilment scripts can now dedupe.