When creating a Six for Six offer, instead of triggering the Six Issues charge on the first paper date and delaying the customer acceptance date to 6 weeks later in order to trigger the quarterly plan as per its catalog definition; I have changed the logic so that we set the Customer Acceptance date as the first paper date (as is normal for all other offer types), as well as triggering the Six Issues plan on that date, and delay the quarterly plan to being triggered on a Specific Date 6 weeks after that.
This change means we no longer have to modify the Customer Acceptance Date - it means what it means for all other offers: the date of first paper and the date of first payment.

I'll double check the fulfilment system is compatible with this change, and will merge if all the billing tests also pass.