Simplify acceptance tests.
We want two scenarios:
1. Unit test suite to run and acceptances tests not to.
2. Acceptance tests to run and unit suite not to.

By providing a single tag `AcceptanceTest`, we can run all tests without it in the first scenario, nd all tests with it in the second scenario.
This also means that when adding a new acceptance test you simply add the same tag and don't have to modify the sbt commands as these are something we don't want to care about during development.

I've removed the individual commands to run individual acceptance tests as these are more boilerplate and unlikely to be used individually given there is a very small number of acceptance tests.