Use the new SalesforceContactServiceUsingZuoraRest in the CheckoutService instead of GetSalesforceContactForSub so we can control whether we are using the buyer or recipient's Salesforce Contact ID.
This also required a change within PurchaserIdentifiers to host both references rather than just the recipient's reference.

Also changed ExactTargetService to use the new SalesforceContactServiceUsingZuoraRest, which has the benefit of not having to make 2 calls to Zuora's getAccount - now we only need to make 1.

Ensured that Gifting is only available on Quarterly and Annual GW plans, not Six for Six.